Modified UpdateRecipe and SaveRecipes function in JsonFileRecipeService.cs
Changed the SaveRecipes function to make it work for string arrays. This version uses JsonSerializer.Serialize to convert the recipe data to a JSON string, then writes that string to the recipes.json file using File.WriteAllText. The JsonSerializerOptions object is configured to use indentation.

*****
The previous version is always giving the formatting error since it would create extra "{" or "]" at the end of the recipes.json to cause bad formats